feat(routes): Add subclass of route to allow for convenient file serving
This does not work because the `filepath` attribute of the `FileGetRoute`
object is not preserved when it is passed to the `register_route` method
of the server, as the parameter type there is `Route` only. The other parameters
are preserved, but later when the object is used as `FileGetRoute` again in
the request handler, the `filepath` attribute is just empty.